perf(props-analyzer): speedup by creating only one ts env
Speedup the props analysis process by creating only
one ts environment for the whole process,
instead of one ts env per component.
